requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic-ai>=0.1.10
asyncio>=3.4.3
pyjwt>=2.10.1
cryptography>=43.0.0